requirements.txt
pandas
flask
flask-login
flask-bcrypt
flask-sqlalchemy
gunicorn
python-binance
